requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.36
redis==5.0.1
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
lxml==5.3.0
python-dateutil==2.8.2
holidays==0.35
pydantic-settings==2.1.0
greenlet==3.1.1